Problem 0033: Use Fraction class for example

+++ Problems.Tests/003X/Problem0033Tests.cs
@@ -2,8 +2,8 @@
 
 namespace Problems.Tests._003X;
 
-public class Problem0033Tests : EulerProblemTestBase<Problem0033, int>
+public class Problem0033Tests : EulerProblemTestBase<Problem0033, long>
 {
-    protected override int ExampleResult => 2;
-    protected override int ProblemResult => 0;
+    protected override long ExampleResult => 2;
+    protected override long ProblemResult => 0;
 }
